+++ .migrations/0002_odd_maria_hill.sql
@@ -1,0 +1,3 @@
+ALTER TABLE "goal_completions" ADD COLUMN "is_archived" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "goals" ADD COLUMN "is_archived" boolean DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE "goals" ADD COLUMN "archived_at" timestamp with time zone;