+++ src/schemas/goals/get-archived-goals-response.ts
@@ -1,0 +1,12 @@
+import { createGoalSchema } from "@/schemas/goals/create-goal";
+import z from "zod";
+
+export const getArchivedGoalsResponseSchema = z.array(
+	createGoalSchema.extend({
+		id: z.cuid2(),
+		isArchived: z.boolean(),
+		archivedAt: z.date().nullable(),
+		createdAt: z.date(),
+		completionCount: z.number(),
+	})
+);